fix: production build was broken since June 6 — site stuck serving stale junk content
Three stacked bugs were preventing every deploy from succeeding, which is
why the live site kept showing deleted/junk numeric-title posts long after
they were removed from the repo:

1. lib/seo.ts had buildHowToJsonLd declared twice (duplicate from a botched
   merge), and app/jobs/[category]/[slug]/page.tsx emitted it twice plus
   called the now-deleted buildFaqJsonLd — all compile errors.

2. components/listings/MobileJobCard.tsx (a client component, via
   PaginatedJobsTable) imported isExpired from lib/content.ts, which pulls
   in Node's `fs`/`path`/gray-matter — breaking the client bundle ("Module
   not found: Can't resolve 'fs'"). Moved the pure date-badge helpers
   (parseDDMMYYYY, isNew, isDeadlineSoon, isExpired) into a new
   lib/dateBadges.ts with zero Node dependencies.

3. mdx_generator.py's YAML helpers didn't escape embedded newlines/
   backslashes from scraped text, producing frontmatter that YAML can't
   parse (a stray "---" on a continuation line reads as a document
   separator → "unexpected end of stream" during static page generation).
   Added _yaml_escape() and fixed 3 existing MDX files corrupted this way;
   deleted one file whose organization field was scraped as literal
   garbage ("Notice Board" with quote marks) and was unrepairable.

Verified with a full `npm run build` — compiles and statically generates
all 2133 job / 1534 result / 277 syllabus pages successfully (first green
build since the June 6 merge).
