+++ content/answer-keys/railway-rpf-constable-answer-key-2025-3003790.mdx
@@ -18,8 +18,7 @@ source: "freejobalert"
 publishedAt: "2026-06-02"
 shortDescription: "the Government has officially announced the Railway Rpf Constable Answer Key 2025 3003790 recruitment 2026 notification. Interested and eligible candidates can apply online for this government government job through the official portal. The last date to apply online for Railway Rpf Constable Answer Key 2025 3003790 is 14-05-2024 (23.59 hours). Do not wait until the deadline — submit your applicati"
 dates: {"Starting Date for Apply Online": "15-04-2024", "Last Date for Apply Online": "14-05-2024 (23.59 hours)", "Dates for Modification window for corrections in application form with payment of modification fee (Please Note: Details filled in ‘Create an Account’ form cannot be modified)": "15-05-2024 to 24-05-2024", "Dates for Access for uploading proper passport size photograph and signature": "15-06-2024 00:00 Hrs to 17-06-2024 23:59 Hrs."}
-fees: {"Physical Standards For PET: Category 1600 meters run 800 meters run Long Jump High Jump Sub-Inspector (Exe) Male 5 min 45 secs -- 14 ft 4 ft Sub-Inspector (Exe) Female 
---- 3 min 40 secs 09 ft 3 ft For PST: Category Height (in CMs) Chest (in CMs) (Only for Male) Male Female Unexpanded Expanded UR/EWS/OBC 165 157 80 85 SC/ST 160 152 76.2 81.2 For Garhwalis, Gorkhas, Marathas, Dogras, Kumaonese and other Categories specified by Govt. 163 155 80 85": "85", "Sub-Inspector (Exe) Female": "3 ft", "UR/EWS/OBC": "85", "SC/ST": "81.2", "Constable (Exe.) Male": "10th Pass or equivalent", "Constable (Exe.) Female": "631", "RPF Constable Result 2025": "Expected in May 2025"}
+fees: {"Physical Standards For PET: Category 1600 meters run 800 meters run Long Jump High Jump Sub-Inspector (Exe) Male 5 min 45 secs -- 14 ft 4 ft Sub-Inspector (Exe) Female \n--- 3 min 40 secs 09 ft 3 ft For PST: Category Height (in CMs) Chest (in CMs) (Only for Male) Male Female Unexpanded Expanded UR/EWS/OBC 165 157 80 85 SC/ST 160 152 76.2 81.2 For Garhwalis, Gorkhas, Marathas, Dogras, Kumaonese and other Categories specified by Govt. 163 155 80 85": "85", "Sub-Inspector (Exe) Female": "3 ft", "UR/EWS/OBC": "85", "SC/ST": "81.2", "Constable (Exe.) Male": "10th Pass or equivalent", "Constable (Exe.) Female": "631", "RPF Constable Result 2025": "Expected in May 2025"}
 importantLinks: [{"label": "RPF Answer Key (24-03-2025 by 6:00 PM) Released", "url": "https://rrb.digialm.com/EForms/loginAction.do?subAction=ViewLoginPage&formId=92912&orgId=33015", "link_type": "answer_key"}, {"label": "Application Status (17-01-2025)", "url": "https://www.rrbapply.gov.in/#/auth/home", "link_type": "other"}, {"label": "Notice for Access for uploading proper passport size photograph and signature", "url": "https://www.rrbmuzaffarpur.gov.in/CommonNotice/660/RPF-Photo-Sign-Re-Upload-_English.pdf", "link_type": "other"}, {"label": "Fee Payment Dates Notice (18-05-2024)", "url": "https://www.rrbbnc.gov.in/notifications/RPF%20CEN%2001%20and%2002%20%20-%20Notice%20for%20unpaid%20applicants.pdf", "link_type": "other"}, {"label": "Correction Window/ Edit Option Link (16-05-2024)", "url": "https://www.rrbapply.gov.in/#/auth/landing", "link_type": "other"}, {"label": "Detailed Notification (15-04-2024)", "url": "https://www.rrbchennai.gov.in/downloads/rpf/cen-02-2024/Final%20notice%20Constable%20RPF%2002-2024_English.pdf", "link_type": "notification"}, {"label": "Notice regarding issue of detailed CEN (20-03-2024)", "url": "https://rrbsecunderabad.gov.in/wp-content/uploads/2024/03/Notice_RPF-CEN-01_2024-RPF-CEN-02_2024.pdf", "link_type": "other"}, {"label": "Official Website", "url": "https://rpf.indianrailways.gov.in/RPF/", "link_type": "official_website"}, {"label": "Download Mobile App", "url": "https://play.google.com/store/apps/details?id=com.freejobalert", "link_type": "other"}, {"label": "Join Telegram Channel", "url": "https://t.me/FreeJobAlertOfficially", "link_type": "other"}, {"label": "Join Whatsapp Channel", "url": "https://www.whatsapp.com/channel/0029VaA1ECQ1noz2PrjLsF3v", "link_type": "other"}]
 ---
