--- content/jobs/government/upsssc-exam-dates-calendar-2026-3052490.mdx
@@ -1,72 +1,0 @@
----
-title: "UPSSSC Exam Calendar 2026 (Out) - Check Schedule & Details"
-slug: "upsssc-exam-dates-calendar-2026-3052490"
-type: "job"
-category: "government"
-dept: ""
-organization: "\"Notice Board\""
-officialWebsite: "https://upsssc.gov.in"
-sourceUrl: "https://www.freejobalert.com/articles/upsssc-exam-dates-calendar-2026-3052490"
-source: "freejobalert"
-publishedAt: "2026-06-04"
-shortDescription: "\"Notice Board\ has officially announced the UPSSSC Exam Calendar 2026 (Out) - Check Schedule & Details recruitment 2026 notification. Interested and eligible candidates can apply online for this government government job through the official portal. Naukri Dhaba brings you the complete UPSSSC Exam Calendar 2026 (Out) - Check Schedule & Details details including eligibility criteria, important dat"
-fees: {"Health Worker (Female) Main Exam": "11 Jan 2026", "Stenographer Main Exam": "20 Jan 2026", "Junior Assistant Main Exam": "01 Feb 2026", "Various Main Exams & Interviews": "Jan to Sep 2026"}
-importantLinks: [{"label": "Download Exam Calendar 2026", "url": "https://upsssc.gov.in/ViewPdf.aspx?oYEmOWynQkSzyPE3pqkE/fxjG2EnU7bWLmOM17w09Iw=", "link_type": "other"}, {"label": "Official Website", "url": "https://upsssc.gov.in", "link_type": "official_website"}]
-howToApply: 
-  - "Go to the official website of \\"Notice Board\ or use the direct apply link on this page at Naukri Dhaba."
-  - "Find the 'UPSSSC Exam Calendar 2026 (Out) - Check Schedule & Details' recruitment notification and click 'Apply Online'."
-  - "Click on 'New Registration' and fill in your name, mobile number, and email ID to create an account."
-  - "Log in with your credentials and fill in Part I of the application — personal details, educational qualifications, and work experience."
-  - "Upload a recent passport-size photograph, your signature, and all required documents in the specified file size and format."
-  - "Pay the application fee online via Debit Card, Credit Card, Net Banking, or UPI."
-  - "Preview your completed application form carefully. Correct any errors before final submission."
-  - "Submit the form and download the confirmation page. Keep a printed copy safe for exam day and document verification."
----
-\"Notice Board\ has officially announced the UPSSSC Exam Calendar 2026 (Out) - Check Schedule & Details recruitment 2026 notification. Interested and eligible candidates can apply online for this government government job through the official portal. Naukri Dhaba brings you the complete UPSSSC Exam Calendar 2026 (Out) - Check Schedule & Details details including eligibility criteria, important dates, application fee, vacancy breakdown, and the direct official apply link. Bookmark this page and check back for admit card and result updates.
-
-## UPSSSC Exam Calendar 2026 (Out) - Check Schedule & Details — Key Details
-
-| Detail | Information |
-|--------|-------------|
-| Organisation | \"Notice Board\ |
-| Job Category | Government |
-
-
-## How to Apply for UPSSSC Exam Calendar 2026 (Out) - Check Schedule & Details
-
-1. Go to the official website of \"Notice Board\ or use the direct apply link on this page at Naukri Dhaba.
-2. Find the 'UPSSSC Exam Calendar 2026 (Out) - Check Schedule & Details' recruitment notification and click 'Apply Online'.
-3. Click on 'New Registration' and fill in your name, mobile number, and email ID to create an account.
-4. Log in with your credentials and fill in Part I of the application — personal details, educational qualifications, and work experience.
-5. Upload a recent passport-size photograph, your signature, and all required documents in the specified file size and format.
-6. Pay the application fee online via Debit Card, Credit Card, Net Banking, or UPI.
-7. Preview your completed application form carefully. Correct any errors before final submission.
-8. Submit the form and download the confirmation page. Keep a printed copy safe for exam day and document verification.
-
-
-## Frequently Asked Questions (FAQ)
-
-### What is the last date to apply for UPSSSC Exam Calendar 2026 (Out) - Check Schedule & Details?
-
-The last date to apply for UPSSSC Exam Calendar 2026 (Out) - Check Schedule & Details is as per official notification. Candidates must submit their online application before this date. No applications are accepted after the deadline under any circumstances.
-
-### How many total vacancies are there in UPSSSC Exam Calendar 2026 (Out) - Check Schedule & Details?
-
-The total vacancy count for UPSSSC Exam Calendar 2026 (Out) - Check Schedule & Details is mentioned in the official notification released by \"Notice Board\. The vacancy breakup by category (General, OBC, SC, ST, EWS) is available in the official notification.
-
-### What is the educational qualification for UPSSSC Exam Calendar 2026 (Out) - Check Schedule & Details?
-
-The minimum educational qualification required for UPSSSC Exam Calendar 2026 (Out) - Check Schedule & Details is . Candidates should verify eligibility from the official notification as requirements may vary by post.
-
-### What is the salary for UPSSSC Exam Calendar 2026 (Out) - Check Schedule & Details?
-
-The salary details for UPSSSC Exam Calendar 2026 (Out) - Check Schedule & Details are mentioned in the official notification by \"Notice Board\. Please refer to the official advertisement for the exact pay scale.
-
-### How to apply for UPSSSC Exam Calendar 2026 (Out) - Check Schedule & Details?
-
-To apply for UPSSSC Exam Calendar 2026 (Out) - Check Schedule & Details: (1) Visit the official website of \"Notice Board\. (2) Find the UPSSSC Exam Calendar 2026 (Out) - Check Schedule & Details notification and click Apply Online. (3) Register, fill the form, upload documents, pay the fee, and submit. (4) Download your application confirmation page. All steps are detailed above on this Naukri Dhaba page.
-
-
----
-*This page is maintained by Naukri Dhaba for informational purposes. All UPSSSC Exam Calendar 2026 (Out) - Check Schedule & Details details are sourced from official \"Notice Board\ notifications. Candidates are advised to verify all information from the official website before applying.*
-
