+++ content/results/government/calcutta-university-result-2025-released-download-pg-course-result-3003131.mdx
@@ -10,8 +10,7 @@ source: "freejobalert"
 publishedAt: "2026-06-04"
 shortDescription: "University Name has officially announced the Calcutta University Result 2025 Released at caluniv.ac.in Direct Link to Download PG Course Result recruitment 2026 notification. Interested and eligible candidates can apply online for this government government job through the official portal. Naukri Dhaba brings you the complete Calcutta University Result 2025 Released at caluniv.ac.in Direct Link to"
 dates: {"Result Status": "Released"}
-importantLinks: [{"label": "Official Site", "url": "http://caluniv.ac.in", "link_type": "other"}, {"label": "Download Mobile App", "url": "https://play.google.com/store/apps/details?id=com.freejobalert", "link_type": "other"}, {"label": "Join Telegram Channel", "url": "https://t.me/FreeJobAlertOfficially", "link_type": "other"}, {"label": "Join WhatsApp Channel", "url": "https://www.whatsapp.com/channel/0029VaA1ECQ1noz2PrjLsF3v", "link_type": "other"}, {"label": "List of successful candidates of PhD Entrance Examination 
---- Department of Psychology", "url": "https://www.caluniv.ac.in/admission/admission.html", "link_type": "other"}]
+importantLinks: [{"label": "Official Site", "url": "http://caluniv.ac.in", "link_type": "other"}, {"label": "Download Mobile App", "url": "https://play.google.com/store/apps/details?id=com.freejobalert", "link_type": "other"}, {"label": "Join Telegram Channel", "url": "https://t.me/FreeJobAlertOfficially", "link_type": "other"}, {"label": "Join WhatsApp Channel", "url": "https://www.whatsapp.com/channel/0029VaA1ECQ1noz2PrjLsF3v", "link_type": "other"}, {"label": "List of successful candidates of PhD Entrance Examination \n--- Department of Psychology", "url": "https://www.caluniv.ac.in/admission/admission.html", "link_type": "other"}]
 ---
 
 Calcutta University Result 2025 is out! Check your Ph.D results now on the official website caluniv.ac.in. Get the direct link to download your Calcutta University Marksheet 2025 here.
